fix: Remove PwService and migrate it to UserService
