+++ api/src/main/java/grooteogi/service/UserService.java
@@ -2,12 +2,16 @@ package grooteogi.service;
 
 import grooteogi.domain.User;
 import grooteogi.dto.LoginDto;
+import grooteogi.dto.PwDto;
 import grooteogi.dto.Token;
 import grooteogi.dto.UserDto;
 import grooteogi.exception.ApiException;
 import grooteogi.exception.ApiExceptionEnum;
 import grooteogi.repository.UserRepository;
 import grooteogi.utils.JwtProvider;
+import grooteogi.utils.Validator;
+import java.sql.Timestamp;
+import java.time.LocalDateTime;
 import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
@@ -25,6 +29,7 @@ public class UserService {
   private final UserRepository userRepository;
   private final JwtProvider jwtProvider;
   private final PasswordEncoder passwordEncoder;
+  private final Validator validator;
 
   public List<User> getAllUser() {
     return userRepository.findAll();
@@ -121,5 +126,21 @@ public class UserService {
 
   public Map refresh(String authorizationHeader, String refreshToken) {
     return jwtProvider.refreshToken(authorizationHeader, refreshToken);
+  }
+
+  public void modifyUserPw(Integer userId, PwDto pwDto) {
+    Optional<User> user = userRepository.findById(userId);
+    if (user.isEmpty()) {
+      throw new ApiException(ApiExceptionEnum.USER_NOT_FOUND_EXCEPTION);
+    }
+
+    if (passwordEncoder.matches(pwDto.getPassword(), user.get().getPassword())) {
+      throw new ApiException(ApiExceptionEnum.BAD_REQUEST_EXCEPTION);
+    }
+
+    validator.confirmPasswordVerification(pwDto.getPassword());
+    user.get().setPassword(passwordEncoder.encode(pwDto.getPassword()));
+    user.get().setModified(Timestamp.valueOf(LocalDateTime.now()));
+    userRepository.save(user.get());
   }
 }
